Updated symmetry number functions to work with new Graph format.
These functions needed quite a bit of work to adjust to storing the bonds
on the atoms itself. In particular, we now need to make deep copies of
the Molecule object in more places, since adding and removing bonds now
modified the Atom objects. This may need more attention in the future,
but seems okay for now.