Fixed bug in cycle detection algorithm caused by new graph format.
The getSmallestSetOfSmallestRings() method generates a copy of the graph
before applying the SSSR algorithm so as to not modify the original.
Before, this was safe because we could make a shallow copy and reuse the
same Vertex and Edge objects, since we didn't store any information about
the graph connectivity on these objects. Now that we are, we must make a
deep copy of the graph to use for the SSSR algorithm. As a result, we
also need to map the vertices of the copy back to those of the original
graph before returning.

This was causing the molecule drawing to fail for cyclic species; the
problem should now be fixed.