Draw reactions vertically aligned on backbone of each molecule.
Aligning to the backbone makes it easier to visually follow the reaction,
as the "paired" reactants and products are more likely to be aligned.
This also accounts for smaller things that might shift the alignment off,
such as radical electrons or subscripts. The only downside is that the
reaction arrow is no longer centered vertically in the generated image.