Better algorithm for finding the "best" straight-chain backbone.
The new algorithm chooses the longest straight-chain with the most
carbon atoms. This ensures a minimum of substituents while still
preferring carbon atoms in the backbone. Two illustrative examples are
isobutanol, which should always have the OH drawn in the backbone, and
acetone, which should never have the C=O drawn in the backbone.

Note that the __findLongestPath() method is now the __findPaths() method
to reflect that it is no longer responsible for returning only the
longest path, but all paths between terminal atoms.