Possibly faster version of Graph.copy().
We spend a lot of time making deep copies of molecules, so any speed boost
here would be valuable. (This is the tradeoff we made in making the
graph isomorphisms themselves faster.) In this commit we eliminate the
expensive calls to list.index(), instead keeping a dict with a mapping of
vertices from the original to the copy.