Faster versions of Atom.copy() and Bond.copy().
The previous implementation was falling back to pure Python. The new
version uses the ExtType.__new__(ExtType) trick to directly instantiate
the extension type object in C. The attributes then need to be populated
manually. This is done inside copy() rather than calling __init__() for
speed reasons.

This change alone should give a 5-10% speed boost, since we spend a lot of
time copying atoms and bonds when generating reactions. Molecule.copy()
should no longer be a significant presence on the CPU profile output.