Hopefully remove redundant spin states when generating reaction templates.
When you removed two radical electrons from an atom type in a functional group
that was {2S,2T} it would create an atom type that was {0,0}.
Hopefully this commit will make it only 0. 

Also, I think it removes a bug whereby previously if you removed 1 radical
electron from something that was 2S, it would become something with a 
single free radical, but a multiplicity of zero! (instead of two).

This should remove the {0,0} part of the RMG-database exporting issue
GreenGroup/RMG-database#18
https://github.com/GreenGroup/RMG-database/issues/18