Fix the atomtypeTest unit tests.
One of the test molecules had impossible valence on an oxygen atom
(it had three bonds). RDKit refused to create this from the smiles sting.
I changed the test molecule, and thus changed the atom numbering in the tests.
Also, fromSMILES() Kekulizes everything, so if we want to test Benzene atom types,
we need to specify things via an Adjacency List.
These tests now pass.