Fix unit test for AtomType which was changing things that shouldn't change.
To test whether the AtomType.setActions() method works, the unit test for
it would change all the properties of the "R" atom type to be those of "Cd".
This then messes up later unit tests that depend on "R" being "R"!!
Specifically, the functional group atom type tests.

We now test the setActions() method on a dummy test atomType.