Added fromOldAdjacencyList method.
It is mostly copied and pasted from before this branch,
but updated to try and create molecules and groups of the new type.
It tries it if the fromAdjacencyList fails.

The hardest part (perhaps not perfect) is how to decide on a molecule
(or group) multiplicity, given that the old style adjacency list stored
these as atomic properties (eg. 2S, 2T, etc).

This commit originally had a bug, that was later fixed:

Fix fromOldAdjacencyList for reading implicit Hydrogens
The constructor for Atoms has changed, and no longer takes
the multiplicity argument.