Perform special check for aromatic isomers to ensure that they are not malformed.
During the updateAtomtypes call, sometimes the aromatic compounds will be
detected as malformed, in one example we found that the indenyl radical produced
an aromatic benzene ring with carbons that had an additional double bond attached.
(ie. 2 double bonds in the 5 member ring adjacent to it.)

This adjlist is still considered aromatic by RDkit but should not be converted:

multiplicity 2
1  C u0 p0 c0 {2,S} {3,S} {4,D}
2  C u0 p0 c0 {1,S} {5,D} {6,S}
3  C u0 p0 c0 {1,S} {8,D} {12,S}
4  C u0 p0 c0 {1,D} {9,S} {13,S}
5  C u0 p0 c0 {2,D} {9,S} {15,S}
6  C u0 p0 c0 {2,S} {7,D} {16,S}
7  C u0 p0 c0 {6,D} {8,S} {10,S}
8  C u0 p0 c0 {3,D} {7,S} {11,S}
9  C u1 p0 c0 {4,S} {5,S} {14,S}
10 H u0 p0 c0 {7,S}
11 H u0 p0 c0 {8,S}
12 H u0 p0 c0 {3,S}
13 H u0 p0 c0 {4,S}
14 H u0 p0 c0 {9,S}
15 H u0 p0 c0 {5,S}
16 H u0 p0 c0 {6,S}

I feel that using a try except block will prevent malformed aromatic adjlists
from being formed but it does not affect other parts of the generateResonanceIsomers
funciton.  However to reduce the slow down of code apply this try except block
only to the new aromatic isomers and not all the new isomers.