create getSymmetryNumber() in rmgpy.molecule.molecule.Molecule
- Rather than initializing the symmetry number to a value that
could represent a possible value (1), set it to the value of -1,
which is impossible.

- getSymmetryNumber first checks whether the result is already cached
before proceeding to calculating it.

- set the return values of calculateSymmetryNumber and getSymmetryNumber
in case of exception to 1 rather than -1, so that an exception in
the algorithm leads to a value that is possibly incorrect but will not
cause any downstream crashes (e.g. ln (-1) ).

- added a test showing the behavior of the getter.