Revert "Fix bug in using rdkit's Chem removeH's function"
This reverts commit 4382640a563bd5ebb0844e59a43afd69a76cfe3e.

I'm not sure what bug this was trying to fix, but it breaks things
by getting RDKit to sanitize molecules when printing SMILES strings.
Sanitizing molecules when printing SMILES may be a nice safety net
to check we don't have dumb molecules, but it also perceives aromaticity,
and prevents us from printing SMILES for Kekule forms, even when our
molecule and the adjacency list is a Kekule form.

The unit test just merged shows this.