Completely overhaul generateAromaticResonanceIsomers
Instead of simply taking the input structure and converting aromatic bonds
to benzene bond types, perform additional analysis in order to obtain
the "most aromatic" form of the molecule. This applies only to cases where
delocalization of a radical into the aromatic ring can break the aromaticity.
In such cases, the algorithm will now generate adjacent resonance structures
and use the ones which give the greatest number of aromatic rings. This means
that in the vast majority of cases, the same resonance structure will be
returned regardless of the input structure. Exceptions include cases where
there are multiple radical positions which give the same number of aromatic
rings, in which case all of the structures are returned.