Fix warning parsing molecules with implicit Hs via RDKit
When running the unit tests for the parser I was seeing the RDKit warning:
> ****
> Pre-condition Violation
> getNumImplicitHs() called without preceding call to calcImplicitValence()
> Violation occurred on line 167 in file
> /opt/RDKit_trunk/Code/GraphMol/Atom.cpp
> Failed Expression: d_implicitValence>-1
> ****

So I googled it, and inspired by this post
https://www.mail-archive.com/rdkit-discuss@lists.sourceforge.net/msg02387.html
I added the UPdatePropertyCache call before calling AddHs.
It seemed to not only make the warning go away but fix the broken SMARTS parsing test,
without making any others fail, so I decided it's worth keeping, and
removed the work_in_progress tag from the now-working test.