Update lone pairs as first step in molecule.updateAtomType()
When lone pairs are left unspecified for groups, we label it as -100 as a placeholder because when we
interface with RDKit, it requires an integer. The new atomTypes require lone pairs to be specified to
something sensible before assigning the specific atomType.
Therefore we have to update lone pairs prior to updating atomtype.

Also removed some instances of updateLonePairs() after updateAtomType() as they are now redundant.