Saturate submols with H in heuristic thermo method for polyrings
Now that we update chargedAtoms when we do Molecule.update(), some functions for the Heuristic ring poly ring corrections are broken. One thing that breaks is that previously correctly atomtyped atoms in submols get charges unless we saturate with Hs. This commit adds H saturation to these cases.

This block of code was added to the functions:

thermo.__addPolyRingCorrectionThermoDataFromHeuristic
thermo.getBicyclicCorrectionThermoDataFromHeuristic