Improved cyclic symmetry algorithm
Previously, the cyclic symmetry algorithm assumed a cycle is planar,
and incorrectly identified some cyclic structures as symmetric
when they are flipped over. This commit looks at the groups attached
to the atom where flipping occurs and makes sure that it would not
result flipping the other groups attached to the ring before assigning
a symmetry value.