changed H bonds to 0.1 in the definitions
A combination of several commits, with these messages:

changed the remove H bonds to remove anything close to 0.1

Hydrogen-bonds now have order 0.1

This way of doing things is a bit fragile, and not very satisfying.
But I think this fixes at least a couple of places where this change
needs to be made.

Detect 0.1 bond orders (for H bonds) more robustly.

Now passes last unit test.
(It was something like 0.1000000000000149)