Allow proper isomorphism w/ isotopes with molecules and groups
This commit modifies the get_element_count method to mention isotope
mass since that leads to a false negative when checking isomorphism
between molecules, which specify elements, and groups, which do not
specify elements.

This commit also adds test methods to isomorphismTest and
isotopesTest, which check that isomorphism works on individual
molecules and within the isotope generation framework.