Implement Graph.getMaxCycleOverlap
Get the maximum number of vertices that are shared between any two cycles
in a graph. For example, if there are only disparate monocycles or no
cycles, the maximum overlap is zero; if there are "spiro" cycles, it is
one; if there are "fused" cycles, it is two; and if there are "bridged"
cycles, it is three.