Sort elements into stable order for chemkin thermo block.
Using the same algorithm as Molecule.get_formula() method
which says iff there is any Carbon then C and H come first,
and everything else is alphabetic.

Before this, it seemed to be unstable, at least with
respect to small changes in RMG, even if it were stable
from one run to another.