A (failing) isomorphism test.
The problem this identifies is that if you give it two different molecules
but the only difference is in a set of atoms that you have already
passed in as a map (eg. because of their labels) then they are thought
to be isomorphic. I.e the initial mapping you pass is not checked.