Bond dissociation energies added for P, F, Br, I
Bond dissociation energies for P, F, Br, and I are added. The source is
added as comment. The bond disccosiation energies in `bde_dict` are
a bit re-arranged. Some are missing because they could not be found
(e.g. N-P, N-Br, N-I, etc).

The information about reference state is also added as comment.