Handle radicals better in conversion from OpenBabel.
The initial pull request in openbabel https://github.com/openbabel/openbabel/pull/1576 sounds 
like it just broke radicals, to be fixed later.
A follow-up https://github.com/openbabel/openbabel/pull/1626 somewhat fixed things but 
"Now reading undervalent atoms from SMILES does not set the spin (it did previously)", 
with a comment "I didn't add an 'op' to set the spin state of undervalent atoms. 
It could be done of course, but the syntax was sufficiently involved (how to specify which 
atoms?, what default atoms?) that I'd rather wait until someone actually has a use case". 

The code introduced there for how to *draw* radicals gives some clues as to how you 
could handle it though, using the expected valence.

That is the approach I take here. The SpinMultiplicity on the OBAtom object is left incorrect
or undefined, and we fix it when converting from_ob_mol to an RMG Molecule.