Fix for when incrementing radical or pair count by more than 1
The logic used to be that if you start with a wildcard number of 
radicals (or pairs) then incrementing it will mean you have a wildcard
that is at least 1 or more. But you could in fact increment by more
than one, using a parameter to the method, so your new wildcard list
shouldn't always start at 1 but should start at however many you incremented by.

Could have done list(range(radical,5)) instead of [1,2,3,4][radical-1:]
but the latter is faster.