Added a work in progress atomtype test: make sample molecule.
Ideally we could make a molecule from every atom type definition
that matches that atom type. 

This is a work in progress test, because we can't.
Two (marked in previous tests) crash. Some others create molecules,
but they don't seem to match the functional group they were designed
to match.