Renaming InChI/SMILES backend option try-all to openbabel-first
try-all is a bit of confusing, while its actual behavior is using openbabel-first whenever possible. Although at rmgpy.molecule.translator line 41-46, there is a try/except to check if openbabel is available, and sometimes only RDKit is included in the BACKEND; Given openbabel is by default installed in the RMG-Py environment, it should be reasonable to call it openbabel-first.