Don't remove H atoms when drawing Hydrogen bonds
This was added in PR #1961 which indroduced
the ability to draw hydrogen bonds.
It was removed without explanation in
ec9515351ea9964a5a8499c1f4066378544eca41
which made drawing work for fragments.
That broke drawing of hydrogen bonds.
The unit test for drawing hydrogen bonds exists,
but just checks that a PDF is made, not
what it looks like. So this wasn't detected.